State Machine Service Plug in
The state machine service plug-in middleware uses the Qt style SCXML format to generate a dynamic state machine. In this function, we can configure the software process and subprocess of the software as we want. The state comes with the attribute configuration function, which can extend our interface style in terms of its functions